refactor(catalog): split DTO god file, deduplicate and remove dead code
- Split common.dto.ts (327 lines, 10 classes) into genre, video, credits,
  user-media-state, availability, and media-base DTOs
- Deduplicate trending item DTOs with TrendingItemBaseDto base class
- Deduplicate verdict DTOs with internal BaseVerdictDto base class
- Extract year validators to presentation/validators/year-range.validator.ts
- Add barrel export dtos/index.ts
- Fix cross-layer import in trending-queries.property.spec.ts
  (was importing from presentation, now from domain)
- Remove unused WatchProviderRegionDto, TrendingMoviesResponseDto
- Remove re-exports of domain constants from DTO files
