+++ apps/api/src/modules/catalog/presentation/dtos/verdict.dto.ts
@@ -8,24 +8,13 @@ import {
   type ShowStatusHintKey,
 } from '../../../shared/verdict';
 
-/**
- * Verdict DTO returned to clients for movie details.
- * Clients use messageKey for i18n lookup.
- */
-export class MovieVerdictDto {
+class BaseVerdictDto {
   @ApiProperty({
     enum: ['warning', 'release', 'quality', 'popularity', 'general'],
     example: 'quality',
     description: 'Verdict type for UI styling',
   })
   type: VerdictType;
-
-  @ApiProperty({
-    example: 'strongRatings',
-    nullable: true,
-    description: 'Message key for i18n lookup on client (details.verdict.movie.*)',
-  })
-  messageKey: MovieVerdictMessageKey;
 
   @ApiProperty({
     example: 'IMDb: 6.3',
@@ -53,15 +42,29 @@ export class MovieVerdictDto {
 }
 
 /**
+ * Verdict DTO returned to clients for movie details.
+ * Clients use messageKey for i18n lookup.
+ */
+export class MovieVerdictDto extends BaseVerdictDto {
+  @ApiProperty({
+    example: 'strongRatings',
+    nullable: true,
+    description: 'Message key for i18n lookup on client (details.verdict.movie.*)',
+  })
+  messageKey: MovieVerdictMessageKey;
+}
+
+/**
  * Show verdict DTO - answers "is it worth it?"
  */
-export class ShowVerdictDto {
+export class ShowVerdictDto extends BaseVerdictDto {
   @ApiProperty({
-    enum: ['warning', 'release', 'quality', 'popularity', 'general'],
-    example: 'quality',
-    description: 'Verdict type for UI styling',
+    example: 'IMDb: 8.2',
+    nullable: true,
+    required: false,
+    description: 'Additional context to display with verdict',
   })
-  type: VerdictType;
+  declare context?: string | null;
 
   @ApiProperty({
     enum: [
@@ -82,30 +85,6 @@ export class ShowVerdictDto {
     description: 'Message key for i18n lookup on client (details.verdict.show.*)',
   })
   messageKey: ShowQualityVerdictKey;
-
-  @ApiProperty({
-    example: 'IMDb: 8.2',
-    nullable: true,
-    required: false,
-    description: 'Additional context to display with verdict',
-  })
-  context?: string | null;
-
-  @ApiProperty({
-    enum: [
-      'newEpisodes',
-      'afterAllEpisodes',
-      'whenOnStreaming',
-      'notifyNewEpisode',
-      'general',
-      'forLater',
-      'notifyRelease',
-      'decideToWatch',
-    ],
-    example: 'forLater',
-    description: 'Hint key for CTA suggestions',
-  })
-  hintKey: VerdictHintKey;
 }
 
 /**
